Fix GLM image patch counter in glm4v source and move regression test
