+++ tests/models/glm4v/test_image_processing_glm4v.py
@@ -252,3 +252,26 @@ class ViTImageProcessingTest(ImageProcessingTestMixin, unittest.TestCase):
             ).pixel_values
             expected_output_image_shape = self.image_processor_tester.expected_output_image_shape(image_inputs)
             self.assertEqual(tuple(encoded_images.shape), expected_output_image_shape)
+
+    def test_get_number_of_image_patches_matches_slow_processor(self):
+        if not is_torchvision_available():
+            self.skipTest("Glm4vImageProcessorFast requires torchvision")
+
+        slow_processor = Glm4vImageProcessor()
+        fast_processor = Glm4vImageProcessorFast()
+
+        test_cases = [
+            (100, 100, {}),
+            (200, 50, {}),
+            (100, 100, {"patch_size": 28}),
+        ]
+
+        for height, width, images_kwargs in test_cases:
+            with self.subTest(height=height, width=width, images_kwargs=images_kwargs):
+                expected = slow_processor.get_number_of_image_patches(
+                    height=height, width=width, images_kwargs=images_kwargs
+                )
+                self.assertEqual(
+                    fast_processor.get_number_of_image_patches(height=height, width=width, images_kwargs=images_kwargs),
+                    expected,
+                )
